requirements.txt
numpy
pandas
tqdm
librosa
scikit-learn
regex
detoxify
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html 
torch==1.11.0+cu113
torchvision==0.12.0+cu113
torchaudio==0.11.0+cu113
transformers
better_profanity
pyannote.audio